Cleaned up database; added database manager and removed database code from MainActivity.
